+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Added a Machine Produce JEI Handler that shows what produces some of the machines
 + Fixed [#40](https://github.com/Buuz135/Industrial-Foregoing/issues/40)
 + Fixed [#19](https://github.com/Buuz135/Industrial-Foregoing/issues/19)
++ Added a Villager Trade Exchanger that trades with villagers automatically
 
 #1.2.1
 + Fixed issue [#37](https://github.com/Buuz135/Industrial-Foregoing/issues/37)